[fix] filter kanban

+++ Se_Capstone_Backend/Capstone.Application/Module/Issues/DTO/IssueDTO.cs
@@ -20,7 +20,7 @@ namespace Capstone.Application.Module.Issues.DTO
         public DateTime? DueDate { get; set; }
         public int Percentage { get; set; } = 0;
         public Priority? Priority { get; set; }
-        public int? EstimatedTime { get; set; }
+        public float? EstimatedTime { get; set; }
         public int? ActualTime { get; set; } = 0;
         public int PercentDone { get; set; } = 0;
         public int Position { get; set; }
@@ -48,7 +48,7 @@ namespace Capstone.Application.Module.Issues.DTO
         public DateTime? DueDate { get; set; }
         public int Percentage { get; set; } = 0;
         public Priority? Priority { get; set; }
-        public int? EstimatedTime { get; set; }
+        public float? EstimatedTime { get; set; }
         public int? ActualTime { get; set; } = 0;
         public int PercentDone { get; set; } = 0;
         public int Position { get; set; }
